Moving User model to Models
